refactor(cost): align reconciliation layers
