+++ src/services/portfolio_transaction_processing_service/app/domain/cost_basis/__init__.py
@@ -6,6 +6,11 @@ from .average_cost_pool_checkpoint import (
     AverageCostPoolRebuildPlan,
     AverageCostPoolTransition,
 )
+from .average_cost_pool_reconciliation import (
+    AverageCostPoolKey,
+    AverageCostPoolReconciliationAssessment,
+    AverageCostPoolReconciliationStatus,
+)
 from .calculation.average_cost_source_allocation import (
     AverageCostPool,
     AverageCostSourceAllocation,
@@ -55,11 +60,6 @@ from .models.effective_fx_rate import EffectiveFxRate
 from .processing_checkpoint import (
     COST_BASIS_STATE_VERSION,
     CostBasisProcessingCheckpoint,
-)
-from .reconciliation import (
-    AverageCostPoolKey,
-    AverageCostPoolReconciliationAssessment,
-    AverageCostPoolReconciliationStatus,
 )
 
 __all__ = [
