+++ src/services/portfolio_transaction_processing_service/app/ports/cost_basis/average_cost_pool_reconciliation.py
@@ -1,14 +1,18 @@
+"""Define the average-cost-pool reconciliation capability port."""
+
 from __future__ import annotations
 
 from typing import Protocol
 
-from ..domain.cost_basis.reconciliation import (
+from ...domain.cost_basis.average_cost_pool_reconciliation import (
     AverageCostPoolKey,
     AverageCostPoolReconciliationAssessment,
 )
 
 
 class AverageCostPoolReconciliationPort(Protocol):
+    """List and reconcile deterministic average-cost-pool candidates."""
+
     async def list_candidates(
         self,
         *,
